# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ReliefMesh API
  version: 0.1.0
  summary: Offline-first, privacy-preserving coordination for local disaster-response exercises and community aid.
  description: |
    REST API of a self-hosted ReliefMesh instance.

    **ReliefMesh is not an emergency service.** For immediate danger to life, health, fire, crime or other
    emergencies, contact the responsible local emergency services. Nothing in this API notifies authorities,
    dispatches responders or performs medical triage.

    ## Authentication
    Sessions use an HTTP-only, `SameSite=Strict` cookie (`__Host-rm_session` over HTTPS, `rm_session` in local
    development). Every state-changing request from an authenticated session must send the CSRF token returned by
    `POST /auth/login` or `GET /auth/session` in the `X-CSRF-Token` header. Browsers must also send a matching
    `Origin`.

    ## Authorization
    All authorization is enforced on the server. Records the caller may not see are reported as `404` so their
    existence is not revealed. Every view contains a `permissions` object computed on the server.

    ## Privacy
    Protected values (contact details and exact locations) are never part of regular responses. They are only
    returned by the `.../protected` endpoints, which record an audit event that is visible to the data subject.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Same origin as the web app
...
security:
  - session: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in with username and password
      description: Rate limited per client IP. Accounts are locked for 15 minutes after 5 consecutive failures. Failure responses are deliberately generic.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              additionalProperties: false
              properties:
                username: { type: string, maxLength: 64 }
                password: { type: string, maxLength: 128, format: password }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: End the current session
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Logged out }
  /api/v1/auth/session:
    get:
      tags: [auth]
      summary: Current session, user and CSRF token
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: Change own password (required after admin-created or reset accounts)
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, minLength: 12, maxLength: 128, format: password }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/me/availability:
    patch:
      tags: [volunteers]
      summary: Set own volunteer availability
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody: { $ref: "#/components/requestBodies/Availability" }
      responses:
        "204": { description: Updated }
        "403": { $ref: "#/components/responses/Error" }
...
  /api/v1/requests:
    get:
      tags: [requests]
      summary: List requests visible to the caller
      description: Requesters see their own requests, volunteers those linked to their assignments, coordinators all requests of their organization. Administrators without a coordination role receive 403. Sensitive descriptions are omitted from list items.
      parameters:
        - { name: status, in: query, schema: { type: string }, description: Comma-separated statuses }
        - { name: category, in: query, schema: { type: string } }
        - { name: urgency, in: query, schema: { $ref: "#/components/schemas/Urgency" } }
        - { name: q, in: query, schema: { type: string, maxLength: 100 }, description: Search reference, title or area (coordinators) }
        - { name: open, in: query, schema: { type: boolean } }
        - { name: mine, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string, enum: [priority, newest, updated] } }
        - { name: updated_since, in: query, schema: { type: string, format: date-time } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200": { description: Page of requests, content: { application/json: { schema: { $ref: "#/components/schemas/RequestList" } } } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [requests]
      summary: Create a request
      description: |
        Urgency is chosen by people, never by the system. Critical urgency requires `emergency_notice_acknowledged: true`.
        For `medicine_pickup`, the title is replaced by a generic logistics title and free-text descriptions are rejected unless enabled by the administrator.
        Sending the same `client_id` again returns the existing request with status 200.
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RequestCreate" } } }
      responses:
        "201": { $ref: "#/components/responses/Request" }
        "200": { $ref: "#/components/responses/Request" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/requests/{id}/notes:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [notes]
      summary: List notes visible to the caller
      responses:
        "200": { $ref: "#/components/responses/NoteList" }
    post:
      tags: [notes]
      summary: Add an append-only note
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody: { $ref: "#/components/requestBodies/Note" }
      responses:
        "201": { $ref: "#/components/responses/Note" }
        "200": { $ref: "#/components/responses/Note" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/offers:
    get:
      tags: [offers]
      summary: List offers visible to the caller
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: allocatable, in: query, schema: { type: boolean } }
        - { name: updated_since, in: query, schema: { type: string, format: date-time } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200": { description: Page of offers, content: { application/json: { schema: { $ref: "#/components/schemas/OfferList" } } } }
    post:
      tags: [offers]
      summary: Create an offer
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/OfferCreate" } } }
      responses:
        "201": { $ref: "#/components/responses/Offer" }
        "200": { $ref: "#/components/responses/Offer" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/offers/{id}/notes:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [notes]
      summary: List offer notes
      responses:
        "200": { $ref: "#/components/responses/NoteList" }
    post:
      tags: [notes]
      summary: Add an offer note
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody: { $ref: "#/components/requestBodies/Note" }
      responses:
        "201": { $ref: "#/components/responses/Note" }
        "200": { $ref: "#/components/responses/Note" }
  /api/v1/assignments:
    get:
      tags: [assignments]
      summary: List assignments (volunteers only ever see their own)
      parameters:
        - { name: request_id, in: query, schema: { type: string, format: uuid } }
        - { name: offer_id, in: query, schema: { type: string, format: uuid } }
        - { name: volunteer_user_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: mine, in: query, schema: { type: boolean } }
        - { name: updated_since, in: query, schema: { type: string, format: date-time } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200": { description: Page of assignments, content: { application/json: { schema: { $ref: "#/components/schemas/AssignmentList" } } } }
    post:
      tags: [assignments]
      summary: Manually create an assignment (coordinators)
      description: Allocating from an offer locks the offer row; over-allocation is rejected with `409 over_allocation` and the remaining quantity, also under concurrent requests. The request must be verified (or already assigned/in progress).
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AssignmentCreate" } } }
      responses:
        "201": { $ref: "#/components/responses/Assignment" }
        "200": { $ref: "#/components/responses/Assignment" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/volunteers:
    get:
      tags: [volunteers]
      summary: Volunteers with availability (coordinators)
      responses:
        "200":
          description: Volunteers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Volunteer" }
        "403": { $ref: "#/components/responses/Error" }
...
  /api/v1/exports/requests.csv:
    get:
      tags: [exports]
      summary: Non-personal request report (organization managers, audited)
      parameters:
        - { name: from, in: query, schema: { type: string }, description: YYYY-MM-DD or RFC 3339 }
        - { name: to, in: query, schema: { type: string } }
      responses:
        "200": { description: CSV, content: { text/csv: { schema: { type: string } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/exports/summary.json:
    get:
      tags: [exports]
      summary: Aggregated summary report (organization managers, audited)
      parameters:
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses:
        "200": { description: Summary, content: { application/json: { schema: { type: object } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/sync/push:
    post:
      tags: [sync]
      summary: Push queued offline operations (idempotent by op_id)
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SyncPush" } } }
      responses:
        "200": { description: Per-operation results, content: { application/json: { schema: { $ref: "#/components/schemas/SyncPushResult" } } } }
  /api/v1/sync/pull:
    get:
      tags: [sync]
      summary: Pull entities for offline caching
      parameters:
        - { name: since, in: query, schema: { type: string, format: date-time }, description: Previous server_time; omit for a full refresh }
      responses:
        "200": { description: Entities, content: { application/json: { schema: { $ref: "#/components/schemas/SyncPull" } } } }
  /api/v1/admin/settings:
    put:
      tags: [admin]
      summary: Update instance settings (partial)
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SettingsUpdate" } } }
      responses:
        "200": { description: Settings, content: { application/json: { schema: { $ref: "#/components/schemas/Settings" } } } }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/categories/{code}:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    patch:
      tags: [admin]
      summary: Relabel, reorder, enable or disable a category
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                label: { type: string, maxLength: 60 }
                description: { type: string, maxLength: 300 }
                enabled: { type: boolean }
                sort_order: { type: integer }
      responses:
        "200": { description: Category, content: { application/json: { schema: { $ref: "#/components/schemas/Category" } } } }
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: List local accounts
      responses:
        "200": { description: Users, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/User" } } } } } } }
    post:
      tags: [admin]
      summary: Create an account with a one-time temporary password
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, display_name, roles]
              properties:
                username: { type: string, pattern: "^[a-z0-9][a-z0-9._@\\-]{2,63}$" }
                display_name: { type: string, maxLength: 80 }
                roles: { type: array, items: { $ref: "#/components/schemas/Role" }, minItems: 1 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  temporary_password: { type: string }
        "422": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/audit:
    get:
      tags: [admin]
      summary: System audit log (newest first)
      parameters:
        - { name: action, in: query, schema: { type: string } }
        - { name: entity_type, in: query, schema: { type: string } }
        - { name: before_id, in: query, schema: { type: integer } }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: Events, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/AuditEvent" } } } } } } }
...
  /api/v1/admin/retention/apply:
    post:
      tags: [admin]
      summary: Apply retention (irreversible redaction of personal data)
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [confirm], properties: { confirm: { type: boolean, const: true } } } } }
      responses:
        "200": { $ref: "#/components/responses/Retention" }
  /api/v1/admin/records/delete:
    post:
      tags: [admin]
      summary: Delete (redact) a record by reference, with reason
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [entity_type, reference, reason]
              properties:
                entity_type: { type: string, enum: [request, offer] }
                reference: { type: string }
                reason: { type: string, minLength: 5, maxLength: 1000 }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/demo/seed:
    post:
      tags: [admin]
      summary: Load a fictional exercise scenario
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [scenario], properties: { scenario: { type: string, enum: [flood, shelter, power-outage] } } } } }
      responses:
        "204": { description: Loaded }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-rm_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
    Offset: { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
    CSRF: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
  requestBodies:
...
    Availability:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [availability]
            properties:
              availability: { type: string, enum: [available, limited, unavailable] }
              availability_note: { type: string, maxLength: 200 }
    Note:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [body]
            properties:
              client_id: { type: string, format: uuid }
              body: { type: string, minLength: 1, maxLength: 4000 }
              visibility: { type: string, enum: [internal, responders, shared] }
              is_sensitive: { type: boolean }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Session:
      description: Session
      content:
        application/json:
          schema:
            type: object
            properties:
              user: { $ref: "#/components/schemas/Me" }
              csrf_token: { type: string }
    Request:
      description: Request
      content: { application/json: { schema: { $ref: "#/components/schemas/Request" } } }
    Offer:
      description: Offer
      content: { application/json: { schema: { $ref: "#/components/schemas/Offer" } } }
    Assignment:
      description: Assignment
      content: { application/json: { schema: { $ref: "#/components/schemas/Assignment" } } }
    Note:
      description: Note
      content: { application/json: { schema: { $ref: "#/components/schemas/Note" } } }
    NoteList:
      description: Notes
      content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/Note" } } } } } }
...
  schemas:
...
    Urgency: { type: string, enum: [critical, high, normal, low] }
...
    Volunteer:
      type: object
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        availability: { type: string, enum: [available, limited, unavailable] }
        availability_note: { type: string }
        active_assignments: { type: integer }
    Category:
      type: object
      properties:
        code: { type: string }
        label: { type: string }
        description: { type: string }
        is_sensitive: { type: boolean }
        enabled: { type: boolean }
        sort_order: { type: integer }
...
    RequestCreate:
      type: object
      additionalProperties: false
      required: [category, urgency]
      properties:
        client_id: { type: string, format: uuid }
        category: { type: string }
        urgency: { $ref: "#/components/schemas/Urgency" }
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 4000 }
        estimated_people_affected: { type: [integer, "null"], minimum: 0 }
        requested_quantity: { type: [integer, "null"], minimum: 0 }
        requested_unit: { type: string, maxLength: 40 }
        requested_by_time: { type: [string, "null"], format: date-time }
        location: { $ref: "#/components/schemas/LocationInput" }
        accessibility_notes: { type: string, maxLength: 1000 }
        contact: { $ref: "#/components/schemas/ContactInput" }
        sensitive_data_flag: { type: boolean }
        requires_formal_authorization: { type: boolean }
        tags: { type: array, items: { type: string }, maxItems: 10 }
        submit: { type: boolean, default: true }
        emergency_notice_acknowledged: { type: boolean }
        client_created_at: { type: string, format: date-time }
...
    RequestList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Request" } }
        has_more: { type: boolean }
...
    OfferCreate:
      type: object
      additionalProperties: false
      required: [category, title, quantity_available]
      properties:
        client_id: { type: string, format: uuid }
        category: { type: string }
        title: { type: string, minLength: 3, maxLength: 120 }
        description: { type: string, maxLength: 4000 }
        quantity_available: { type: integer, minimum: 0 }
        unit: { type: string, maxLength: 40 }
        availability_start: { type: [string, "null"], format: date-time }
        availability_end: { type: [string, "null"], format: date-time }
        pickup_or_delivery_mode: { type: string, enum: [pickup, delivery, either, on_site] }
        location: { $ref: "#/components/schemas/LocationInput" }
        accessibility_notes: { type: string, maxLength: 1000 }
        restrictions: { type: string, maxLength: 1000 }
        contact: { $ref: "#/components/schemas/ContactInput" }
        tags: { type: array, items: { type: string } }
        publish: { type: boolean, default: true }
        client_created_at: { type: string, format: date-time }
...
    OfferList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Offer" } }
        has_more: { type: boolean }
    AssignmentCreate:
      type: object
      additionalProperties: false
      required: [request_id]
      properties:
        client_id: { type: string, format: uuid }
        request_id: { type: string, format: uuid }
        offer_id: { type: [string, "null"], format: uuid }
        volunteer_user_id: { type: [string, "null"], format: uuid }
        team_label: { type: string, maxLength: 120 }
        quantity_assigned: { type: integer, minimum: 0 }
        unit: { type: string }
        instructions: { type: string, maxLength: 2000, description: Free-text operational notes only. No routing, no safety claims. }
        protected_contact_access_granted: { type: boolean }
        pickup_location_access_granted: { type: boolean }
        destination_location_access_granted: { type: boolean }
        eta_text: { type: string, maxLength: 120 }
...
    AssignmentList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Assignment" } }
        has_more: { type: boolean }
...
    SyncPush:
      type: object
      required: [operations]
      properties:
        operations: { type: array, maxItems: 200, items: { $ref: "#/components/schemas/SyncOperation" } }
    SyncPushResult:
      type: object
      properties:
        server_time: { type: string, format: date-time }
        results:
          type: array
          items:
            type: object
            properties:
              op_id: { type: string, format: uuid }
              status: { type: string, enum: [applied, duplicate, conflict, rejected, retry] }
              entity_type: { type: string }
              entity_id: { type: string, format: uuid }
              reference: { type: string }
              error: { type: object }
              entity: { type: object, description: Current server state (for conflicts and applied operations) }
    SyncPull:
      type: object
      properties:
        server_time: { type: string, format: date-time }
        full: { type: boolean }
        has_more: { type: boolean }
        settings: { $ref: "#/components/schemas/ClientSettings" }
        requests: { type: array, items: { $ref: "#/components/schemas/Request" } }
        offers: { type: array, items: { $ref: "#/components/schemas/Offer" } }
        assignments: { type: array, items: { $ref: "#/components/schemas/Assignment" } }